% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desraj.R
\name{desraj}
\alias{desraj}
\title{Desraj Ordered Estimator}
\usage{
desraj(y, p)
}
\arguments{
\item{y}{vector of vaues of sampled units as per the order of selection}

\item{p}{vector of selection probabilities as per the order of selection}
}
\description{
Desraj estimator computes the estimated value of a finite population
              total when values of the study variable for the sampled units
              and the corresponding selection probabilities are given as per the order
              of selection. It is meant for use in Probability Proportional to Size
              Without Replacement Sampling Scheme.
}
\examples{
 y<-c(16,13,12,10)
 p<-c(0.21, 0.34, 0.12,0.10)
 desraj(y,p)
}
\references{
Sampath,S.(2005) Sampling Theory and Methods,Alpha Science International, Ltd
}
